Password updates
Changed the way passwords are used in the system. Input is first
hashed with SHA256 then the first 16 bytes are taken. Perhaps in
future, I should add all the bytes every 16 together mod 255?

Also used the password to generate a seed to permute the files
before use. Adds a little bit for randomness into the system
instead of always using alphabetical order.