Write Blocking
Added a write blocking feature to stop accidental changes to the filesystem.

The threat model for this feature is being unable to trust a cloud storage solution. If your storage provider goes rogue or is pressured by a government to take snapshots of your data, the constantly changing hashes of pictures (while the image is staying the same) is an easy red flag. Upon closer inspection, seeing that it's the LSBs of the DCTs changing would be a huge red flag.

I wanted to help prevent this from happening so the idea is that you create the system locally and copy your data into there, then upload those files. As long as you always access it with the writeblocker on, the file hash should never change and the flags are gone.

The write blocker functions at the StorageManager level. Writes will return -EPERM (but the file manager currently ignores this), flushes will silently return while doing nothing. It's worth noting that since the FileManager ignores this, any client applications have no idea their write failed and the data is actually saved in the memory buffer.