AES first block corruption fix + JPEG granularity
Fixed a bug where under certain cicumstances, the wrong IV would be used for file decryption resulting in the loss of the first 16 bytes of the file.

Added the ability to easily balance data storage and JPEG alterations. The updates for this can be found in the header file Jpeg.h